This patch changes unsigned [int] to u16 for subsystem IDs.  They're in the hardware and have a specific size.
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>



git-svn-id: svn://coreboot.org/repository/coreboot-v3@1069 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ device/pci_device.c
@@ -710,8 +710,7 @@ void pci_bus_reset(struct bus *bus)
 	delay(1);
 }
 
-void pci_dev_set_subsystem(struct device *dev, unsigned int vendor,
-			   unsigned int device)
+void pci_dev_set_subsystem(struct device *dev, u16 vendor, u16 device)
 {
 	pci_write_config32(dev, PCI_SUBSYSTEM_VENDOR_ID,
 			   ((device & 0xffff) << 16) | (vendor & 0xffff));
