+++ southbridge/nvidia/mcp55/mcp55.h
@@ -24,8 +24,7 @@
 
 #define PCI_MCP55_SUBSYSTEM_VENDOR_ID 0x40
 
-void mcp55_pci_dev_set_subsystem(struct device *dev, unsigned int vendor,
-			   unsigned int device);
+void mcp55_pci_dev_set_subsystem(struct device *dev, u16 vendor, u16 device);
 extern struct pci_operations mcp55_pci_dev_ops_pci;
 
 #endif /* MCP55_H */
